git subrepo pull kigs
subrepo:
  subdir:   "kigs"
  merged:   "0e151b906f"
upstream:
  origin:   "https://github.com/assoria/kigs.git"
  branch:   "master"
  commit:   "0e151b906f"
git-subrepo:
  version:  "0.4.1"
  origin:   "???"
  commit:   "???"

Former-commit-id: 26548b584836072f7c1d69fb69655e3354759476
Former-commit-id: 21f9058c94aded23d4ba53d030119f400d5a9166

+++ framework/2DLayers/Sources/UI/UIGridLayout.cpp
@@ -91,7 +91,7 @@ void UIGridLayout::RecomputeLayout()
 
 
 		int r = 0;
-		unsigned int columns = mColumns == 0 ? items.size() : mColumns;
+		unsigned int columns = mColumns == 0 ? (unsigned int)items.size() : mColumns;
 		while (it != end) {
 
 			float widthAccumulator = 0;
